Drone Fleet Map Implementation
The app now has a start activity where setting can be launched or a
flight path may be created.

The setting screen displays the IP addresses for two physical drones for
live demonstration as well as specify the IP address for the simulator
and the number of drones to simulate in flight.

The map screen allows the user to create either a line or shape flight
path that the drone(s) will follow as a mission. From here an arraylist
is created that stores the latitude-longitude coordinates of the points
along the shape that will be passed on to the rest of the program. Upon
completion the Go command launches a screen where the user will select
either simulation or live drones for execution. At this point the user
is returned to the map screen to watch the results as they occur.